Fork from ipa320, large re-write and re-structuring.
This is the beginning (though mostly finished) of a massive re-write,
mostly because there were far too many either issues in the
implementation of the spec, chaotic to find actual files, or missing
proper documentation.
As I am now actively developing with VDA5050, this repo will also be
actively maintained, if for nothing else than professional use.

Signed-off-by: JWCS <sauve031@umn.edu>